still working on fixing seg fault

+++ AVLTree.h
@@ -48,7 +48,7 @@ public:
     size_t size() const;
 
     // Returns AVL tree height
-    size_t getHeight() const;
+    int getHeight() const;
 
     //helper to get subtree height
     int getSubnodeHeight(BSTNode* node) const;
